Moving methods to class
